+++ src/style.css
@@ -99,10 +99,10 @@ body {
   font-size: 14px;
 }
 
-.task-desc{
+.task-desc {
   outline: none;
 }
 
-.displaynt{
+.displaynt {
   display: none;
 }
