+++ src/task.js
@@ -1,0 +1,7 @@
+export default class Task {
+  constructor(index, desc, completed) {
+    this.index = index;
+    this.desc = desc;
+    this.completed = completed;
+  }
+}
